+++ src/dftbp/dftbplus/initprogram.F90
@@ -151,9 +151,9 @@ module dftbp_dftbplus_initprogram
 #:endif
 #:if WITH_TRANSPORT
   use dftbp_dftbplus_inputdata, only : TNEGFInfo
-  use dftbp_transport_negfint, only : TNegfInt, TNegfInt_init
-  use dftbp_transport_negfvars, only : TTransPar
+  use dftbp_transport_negfint, only : TNegfInt, TNegfInt_init, transportPeriodicSetup
 #:endif
+  use dftbp_transport_negfvars, only : TTransPar
   implicit none
 
   private
@@ -233,6 +233,9 @@ module dftbp_dftbplus_initprogram
     !> Are atomic coordinates fractional?
     logical :: tFracCoord
 
+    !> If the extended structure outside of the central cell be outputed, name of file
+    character(lc) :: extendedGeomFile = ""
+
     !> Tolerance for SCC cycle
     real(dp) :: sccTol
 
@@ -251,7 +254,7 @@ module dftbp_dftbplus_initprogram
     !> Normalized vectors in those directions
     real(dp) :: normOrigLatVec(3,3)
 
-    !> Reciprocal vectors in 2 pi units
+    !> Reciprocal vectors in 2 pi units / inverse of the lattice vector matrix
     real(dp), allocatable :: invLatVec(:,:)
 
     !> Cell volume
@@ -1046,9 +1049,9 @@ module dftbp_dftbplus_initprogram
 
   #:if WITH_TRANSPORT
     !> Transport variables
-    type(TTransPar) :: transpar
     type(TNEGFInfo) :: ginfo
   #:endif
+    type(TTransPar) :: transpar
 
     !> Transport interface (may be dummy placeholder, if built without transport)
     type(TNegfInt) :: negfInt
@@ -1398,7 +1401,7 @@ contains
     call initGeometry_(input, this%nAtom, this%nType, this%tPeriodic, this%tHelical,&
         & this%boundaryCond, this%coord0, this%species0, this%tCoordsChanged, this%tLatticeChanged,&
         & this%latVec, this%origin, this%recVec, this%invLatVec, this%cellVol, this%recCellVol,&
-        & errStatus)
+        & input%transpar, errStatus)
     if (errStatus%hasError()) call error(errStatus%message)
 
     ! Get species names and output file
@@ -1555,6 +1558,14 @@ contains
     end if
     this%tFracCoord = input%geom%tFracCoord
 
+    if (input%ctrl%areAllAtomsPrinted .and. isIoProc) then
+      if (len(trim(this%geoOutFile)) > 0) then
+        this%extendedGeomFile = "extendedGeom_"//trim(this%geoOutFile)//".xyz"
+      else
+        this%extendedGeomFile = "extendedGeom.xyz"
+      end if
+    end if
+
     ! no point if not SCC
     this%isSccConvRequired = (input%ctrl%isSccConvRequired .and. this%tSccCalc)
 
@@ -1688,6 +1699,9 @@ contains
     if (this%tUpload) then
       call initUploadArrays_(input%transpar, this%orb, this%nSpin, this%tMixBlockCharges,&
           & this%shiftPerLUp, this%chargeUp, this%blockUp)
+      if (input%transpar%ncont < 1) then
+        call error("At least one contact is required for an UploadContacts task")
+      end if
     end if
     call initTransport_(this, env, input, this%electronicSolver, this%nSpin, this%tempElec,&
         & this%tNegf, this%isAContactCalc, this%mu, this%negfInt, this%ginfo, this%transpar,&
@@ -2238,10 +2252,10 @@ contains
       end if
       if (allocated(input%ctrl%dispInp%slakirk)) then
         allocate(slaKirk)
-        if (this%tPeriodic) then
+        if (this%tPeriodic .and. this%transpar%nCont == 0) then
           call DispSlaKirk_init(slaKirk, input%ctrl%dispInp%slakirk, this%latVec)
         else if (this%tHelical) then
-          call error("Slater-Kirkwood incompatible with helical boundary conditions")
+          call error("Slater-Kirkwood currently incompatible with helical boundary conditions")
         else
           call DispSlaKirk_init(slaKirk, input%ctrl%dispInp%slakirk)
         end if
@@ -2249,7 +2263,7 @@ contains
 
       elseif (allocated(input%ctrl%dispInp%uff)) then
         allocate(uff)
-        if (this%tPeriodic) then
+        if (this%tPeriodic .and. this%transpar%nCont == 0) then
           call DispUff_init(uff, input%ctrl%dispInp%uff, this%nAtom, this%species0, this%latVec)
         else
           call DispUff_init(uff, input%ctrl%dispInp%uff, this%nAtom)
@@ -2260,7 +2274,7 @@ contains
         block
           type(TSDFTD3), allocatable :: dftd3
           allocate(dftd3)
-          if (this%tPeriodic) then
+          if (this%tPeriodic .and. this%transpar%nCont == 0) then
             call TSDFTD3_init(dftd3, input%ctrl%dispInp%dftd3, this%nAtom, this%species0, &
                 & this%speciesName, this%coord0, this%latVec)
           else
@@ -2272,13 +2286,14 @@ contains
 
       else if (allocated(input%ctrl%dispInp%sdftd3)) then
         allocate(sdftd3)
-        if (this%tPeriodic) then
+        if (this%tPeriodic .and. this%transpar%nCont == 0) then
           call init(sdftd3, input%ctrl%dispInp%sdftd3, this%nAtom, this%species0, this%speciesName,&
               & this%latVec)
         else
           call init(sdftd3, input%ctrl%dispInp%sdftd3, this%nAtom, this%species0, this%speciesName)
         end if
         call move_alloc(sdftd3, this%dispersion)
+
       else if (allocated(input%ctrl%dispInp%dftd4)) then
         allocate(dftd4)
         if (allocated(this%reks)) then
@@ -2287,12 +2302,16 @@ contains
                 & force calculation in REKS")
           end if
         end if
+        if (this%transpar%nCont /= 0) then
+          call error("DFTD4 model not currently supported for transport calculations")
+        end if
         if (this%tPeriodic) then
           call init(dftd4, input%ctrl%dispInp%dftd4, this%nAtom, this%speciesName, this%latVec)
         else
           call init(dftd4, input%ctrl%dispInp%dftd4, this%nAtom, this%speciesName)
         end if
         call move_alloc(dftd4, this%dispersion)
+
     #:if WITH_MBD
       else if (allocated(input%ctrl%dispInp%mbd)) then
         if (this%isLinResp) then
@@ -2303,6 +2322,8 @@ contains
             call error("Calculation of self-consistent MBD/TS is not currently compatible with&
                 & force calculation in REKS")
           end if
+        else if (this%transpar%nCont /= 0) then
+          call error("MBD model not currently supported for transport calculations")
         end if
         allocate (mbd)
         associate (inp => input%ctrl%dispInp%mbd)
@@ -2321,8 +2342,11 @@ contains
               & which may result in long gradient calculation times for large systems")
         end if
     #:endif
+
       end if
+
       this%cutOff%mCutOff = max(this%cutOff%mCutOff, this%dispersion%getRCutOff())
+
     end if
 
     this%areSolventNeighboursSym = .false.
@@ -2999,9 +3023,6 @@ contains
   #:endif
 
     if (this%tNegf) then
-      if (allocated(this%dispersion)) then
-        call error("Dispersion not currently available with transport calculations")
-      end if
       if (this%isLinResp) then
         call error("Linear response is not compatible with transport calculations")
       end if
@@ -5151,6 +5172,9 @@ contains
       call clearFile(trim(this%geoOutFile) // ".gen")
       call clearFile(trim(this%geoOutFile) // ".xyz")
     end if
+    if (len(trim(this%extendedGeomFile)) > 0) then
+      call clearFile(trim(this%extendedGeomFile))
+    end if
     if (allocated(this%electrostatPot)) then
       call clearFile(this%electrostatPot%espOutFile)
     end if
@@ -6760,15 +6784,15 @@ contains
   !> Initializes the variables directly related to the user specified geometry.
   subroutine initGeometry_(input, nAtom, nType, tPeriodic, tHelical, boundaryCond, coord0,&
       & species0, tCoordsChanged, tLatticeChanged, latVec, origin, recVec, invLatVec, cellVol,&
-      & recCellVol, errStatus)
+      & recCellVol, transpar, errStatus)
 
-    !> Geometry input
+    !> Input variables to use for setput
     type(TInputData), intent(in) :: input
 
-    !> Total number of atoms, including images
+    !> Number of unique atoms in the system
     integer, intent(out) :: nAtom
 
-    !> Number of chemical types
+    !> Number of chemical types of atoms
     integer, intent(out) :: nType
 
     !> Is this a periodic geometry
@@ -6777,13 +6801,13 @@ contains
     !> Is this a helical geometry
     logical, intent(out) :: tHelical
 
-    !> Boundary conditions
+    !> Boundary conditions on the calculation
     type(TBoundaryConditions), intent(out) :: boundaryCond
 
-    !> Central cell coordinates
+    !> Coordinates of the central cell atoms
     real(dp), allocatable, intent(out) :: coord0(:,:)
 
-    !> Central cell chemical species
+    !> Species of the central cell atoms
     integer, allocatable, intent(out) :: species0(:)
 
     !> Have the coordinates been updated without updating dependent variables
@@ -6810,8 +6834,16 @@ contains
     !> Volume of the reciprocal space unit cell
     real(dp), intent(out) :: recCellVol
 
+    !> Transport calculation parameters
+    type(TTransPar), intent(in) :: transpar
+
     !> Operation status, if an error needs to be returned
     type(TStatus), intent(inout) :: errStatus
+
+    real(dp), allocatable :: tmpCoords0(:,:)
+    integer, allocatable :: tmpSpecies0(:), nExtraContAtoms(:)
+    integer :: iCont, nContAts, iStart, iEnd, iStart2, iStructOffSet, iAt
+    real(dp) :: contactVector(3)
 
     nAtom = input%geom%nAtom
     nType = input%geom%nSpecies
@@ -6828,39 +6860,72 @@ contains
     end if
     @:PROPAGATE_ERROR(errStatus)
 
+  #:if WITH_TRANSPORT
+    if (tPeriodic) then
+       call transportPeriodicSetup(input%transpar, input%geom%latVecs, boundaryCond)
+    end if
+  #:endif
+
     coord0 = input%geom%coords
     species0 = input%geom%species
+
+    if (transpar%ncont > 0) then
+      ! Extend contact regions for dispersion interaction distance
+      allocate(nExtraContAtoms(transpar%ncont), source=0)
+      do iCont = 1, transpar%ncont
+        nExtraContAtoms(iCont) = transpar%contacts(iCont)%idxrange(2) + 1&
+            & - transpar%contacts(iCont)%idxrange(1)
+      end do
+      allocate(tmpCoords0(3, nAtom + sum(nExtraContAtoms)))
+      allocate(tmpSpecies0(nAtom + sum(nExtraContAtoms)))
+      tmpCoords0(:, :nAtom) = coord0
+      tmpSpecies0(:nAtom) = species0
+      iStructOffSet = nAtom
+      do iCont = 1, transpar%ncont
+        iStart = transpar%contacts(iCont)%idxrange(1)
+        iEnd = transpar%contacts(iCont)%idxrange(2)
+        iStart2 = iStart + (iEnd - iStart + 1) / 2
+        ! Vector pointing into contact, away from device:
+        contactVector(:) = 2.0_dp * (coord0(:,iStart2) - coord0(:,iStart))
+        tmpSpecies0(iStructOffSet+1:iStructOffSet+nExtraContAtoms(iCont)) =&
+            & species0(transpar%contacts(iCont)%idxrange(1) : transpar%contacts(iCont)%idxrange(2))
+        tmpCoords0(:, iStructOffSet+1:iStructOffSet+nExtraContAtoms(iCont)) =&
+            & coord0(:,transpar%contacts(iCont)%idxrange(1):transpar%contacts(iCont)%idxrange(2))
+        do iAt = iStructOffSet+1, iStructOffSet+nExtraContAtoms(iCont)
+          tmpCoords0(:, iAt) = tmpCoords0(:, iAt) + contactVector
+        end do
+        iStructOffSet = iStructOffSet + nExtraContAtoms(iCont)
+      end do
+      block
+        use dftbp_io_formatout, only : writeXYZFormat
+        call writeXYZFormat("tmp.xyz", tmpCoords0, tmpSpecies0, input%geom%speciesNames)
+      end block
+      deallocate(tmpCoords0)
+      deallocate(tmpSpecies0)
+    end if
+
     tCoordsChanged = .true.
 
-    cellVol = 0.0_dp
-    recCellVol = 0.0_dp
+    tLatticeChanged = .false.
     if (tPeriodic) then
       tLatticeChanged = .true.
       latVec = input%geom%latVecs
       origin = input%geom%origin
       allocate(recVec(3, 3))
       allocate(invLatVec(3, 3))
-      invLatVec(:,:) = latVec
-      call invert33(invLatVec)
-      invLatVec = reshape(invLatVec, [3, 3], order=[2, 1])
-      recVec = 2.0_dp * pi * invLatVec
-      cellVol = abs(determinant33(latVec))
-      recCellVol = abs(determinant33(recVec))
-    elseif (tHelical) then
+    else if (tHelical) then
       origin = input%geom%origin
       latVec = input%geom%latVecs
-      allocate(recVec(1, 1))
-      recVec = 1.0_dp / latVec(1,1)
+      allocate(recVec(0, 0))
       allocate(invLatVec(0, 0))
     else
       allocate(latVec(0, 0))
       allocate(origin(0))
       allocate(recVec(0, 0))
       allocate(invLatVec(0, 0))
-      cellVol = 0.0_dp
-      recCellVol = 0.0_dp
-      tLatticeChanged = .false.
     end if
+
+    call boundaryCond%handleBoundaryChanges(latVec, invLatVec, recVec, cellVol, recCellVol)
 
   end subroutine initGeometry_